[libromdata] GcnPartition::partition_size_used(): Cast the FST offset+size to int64_t before shifting.
This shouldn't cause any problems with original GCN/Wii disc images.

A malicious GCN/Wii disc image could cause some shenanigans here due to
32-bit wraparound, but it's unlikely that it'd cause any real problems.

This fixes Coverity#1398404.

+++ src/libromdata/disc/GcnPartition.cpp
@@ -183,7 +183,7 @@ int64_t GcnPartition::partition_size_used(void) const
 	}
 
 	// FST offset and size.
-	int64_t size = (d->bootBlock.fst_offset + d->bootBlock.fst_size) << d->offsetShift;
+	int64_t size = (int64_t)(d->bootBlock.fst_offset + d->bootBlock.fst_size) << d->offsetShift;
 	
 	// Get the FST used size.
 	size += d->fst->totalUsedSize();
